Set style in rocket items
