+++ src/components/Rockets/Rockets.css
@@ -1,0 +1,12 @@
+.rocket-img {
+  width: 50vh;
+  height: 35vh;
+}
+
+.ml-3 {
+  margin-left: 3vh;
+}
+
+.font-w h4 {
+  font-weight: 600;
+}
